requirements.txt
# Ultra-minimal requirements for Railway deployment
# Only essential dependencies to avoid transformers import

# Web framework
Flask>=2.3.0
Flask-CORS>=4.0.0
gunicorn>=21.2.0

# Core essentials  
openai>=1.0.0
python-dotenv
pickledb
dacite

# Basic utilities (very lightweight)
requests
urllib3